refactor: rename and move localTimeInput directive
to its own file
