+++ client/app/components/group/_pickupEditCreate/pickupEditCreate.html
@@ -38,7 +38,7 @@
           <label translate="CREATEPICKUP.TIME"></label>
           <md-icon class="fa fa-clock-o fa-2x"></md-icon>
           <input ng-model="$ctrl.time" name="time" placeholder="20:00"
-                 type="time" custom-format-hack required />
+                 type="time" local-time-input required />
           <div ng-messages="pickupEditCreateForm.time.$error" role="alert">
             <span ng-message="time" translate="CREATEPICKUP.TIME_ERROR"></span>
           </div>
